# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Books API - OpenAPI 3.0",
    "description": "Books API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
...
  "paths": {
...
    "/books/borrow": {
      "post": {
        "tags": ["book"],
        "description": "Borrow Book",
        "operationId": "borrowBook",
        "requestBody": {
          "description": "Borrow a Book by an Member",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BorrowBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BorrowBadRequest"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BorrowBookRequest": {
        "type": "object",
        "properties": {
          "bookId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "memberId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "qty": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
...
      "BorrowBadRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "msg": {
            "type": "string",
            "example": "Book Harry Potter already borrowed."
          },
          "param": {
            "type": "string",
            "example": "bookId"
          },
          "location": {
            "type": "string",
            "example": "body"
          }
        }
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "code": {
            "type": "string",
            "example": "B001"
          },
          "title": {
            "type": "string",
            "example": "A Study in Scarlet"
          },
          "author": {
            "type": "string",
            "example": "Arthur Conan Doyle"
          },
          "stock": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "example": "2023-03-28T15:40:54.290Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2023-03-28T15:40:54.290Z"
          }
        }
      },
...
    }
  }
}